source/api/dataplane/infrastructure/requirements.txt
aws-cdk.core>=1.85.0,<2.0
aws-cdk.aws-dynamodb>=1.85.0,<2.0
aws-cdk.aws-iam>=1.85.0,<2.0
aws-cdk.aws-lambda>=1.85.0,<2.0
aws-cdk.aws-lambda-event-sources>=1.85.0,<2.0
aws-cdk.aws-sqs>=1.85.0,<2.0
aws-cdk.aws-ssm>=1.85.0,<2.0
chalice[cdk]
